Create plotting_from_csv_and_curvefitting_to_capacitance
This code is to be used after you've run the data collection code to gather you values at a steady distance. The code averages the value of distance from the MPR sensor to plot as a continuous line. The code also begins its curve fitting, similar to poly-fit so you can see how you sensor lines up with the graph of capacitance.